Communication folder refactor
